Add admin endpoint to see al active users and their sessions

+++ src/main/java/com/aboher/sessionsecureapp/service/impl/SessionService.java
@@ -9,6 +9,7 @@ import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.session.FindByIndexNameSessionRepository;
 import org.springframework.session.Session;
@@ -29,6 +30,7 @@ public class SessionService {
     private final HttpServletRequest request;
     private final HttpSession httpSession;
     private final FindByIndexNameSessionRepository<? extends Session> sessionRepository;
+    private final RedisService redisService;
 
     public HttpSession invalidateOldSessionAndCreateNewOne() {
         HttpSession oldSession = request.getSession(false);
@@ -39,7 +41,10 @@ public class SessionService {
     }
 
     public void invalidateSession(String id) {
-        sessionRepository.deleteById(id);
+        Session session = getSessionBySessionId(id);
+        if (sessionBelongsToCurrentUser(session) || getRoles().contains(Role.ROLE_ADMIN)) {
+            sessionRepository.deleteById(id);
+        }
     }
 
     public Date getExpirationDate() {
@@ -107,17 +112,47 @@ public class SessionService {
     }
 
     public Set<String> getAllActiveSessionsIds(Principal principal) {
-        return sessionRepository.findByPrincipalName(principal.getName()).values().stream()
+        return getAllActiveSessionsIds(principal.getName());
+    }
+
+    public Set<String> getAllActiveSessionsIds(String principalName) {
+        return sessionRepository.findByPrincipalName(principalName).values().stream()
                 .map(Session::getId)
                 .collect(Collectors.toSet());
     }
 
+    public Set<String> getAllActiveSessionsPrincipalNames() {
+        return redisService.getPrincipalNamesOfAllTheUsersWithActiveSessions();
+    }
+
     public SessionInfo getSessionInfoBySessionId(String id) {
+        Session session = getSessionBySessionId(id);
+        // I check this because a hacker could use this endpoint to get SessionInfo
+        // from other users by guessing session ids, and I don't want to allow that.
+        // So, if a User with role USER, tries to get a session that doesn't belong
+        // to him, he will get the same message as if it didn't exist at all.
+        if (!sessionBelongsToCurrentUser(session) && !getRoles().contains(Role.ROLE_ADMIN)) {
+            throw new InvalidSessionException("There is no session with id = " + id);
+        }
+        return getSessionInfo(session);
+    }
+
+    private boolean sessionBelongsToCurrentUser(Session session) {
+        String currentEmail = getEmail();
+        SecurityContext securityContext = session.getAttribute("SPRING_SECURITY_CONTEXT");
+        Authentication authentication = securityContext.getAuthentication();
+        String sessionEmail = authentication.getName();
+
+//        String sessionEmail = session.getAttribute(FindByIndexNameSessionRepository.PRINCIPAL_NAME_INDEX_NAME);
+        return currentEmail.equals(sessionEmail);
+    }
+
+    public Session getSessionBySessionId(String id) {
         Session session = sessionRepository.findById(id);
         if (session == null) {
             throw new InvalidSessionException("There is no session with id = " + id);
         }
-        return getSessionInfo(session);
+        return session;
     }
 
     public SessionInfo getSessionInfo(Session session) {
